requirements.txt
python-dotenv==1.0.0
Django==4.1.4
djangorestframework==3.14.0
requests==2.26.0
djangorestframework-simplejwt==4.7.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
